backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pybit==5.11.0
python-dotenv==1.0.1
pydantic-settings==2.6.1


